角色管理和菜单管理

+++ projectroot/pms/src/main/java/com/xxx/pms/entity/UserFavoriteMenu.java
@@ -42,6 +42,6 @@ public class UserFavoriteMenu {
      *
      *   排序
      */
-    private Byte sort;
+    private Integer sort;
     
 }
